+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/index.js":"1","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/landing-page.js":"2","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/workouts-list-sw.js":"3","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/services/token-service-lf.js":"4","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/services/auth-api-service-lf.js":"5","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/header.js":"6"},{"size":358,"mtime":1607133044997,"results":"7","hashOfConfig":"8"},{"size":918,"mtime":1607133044998,"results":"9","hashOfConfig":"8"},{"size":2009,"mtime":1607794837647,"results":"10","hashOfConfig":"8"},{"size":1707,"mtime":1607133044999,"results":"11","hashOfConfig":"8"},{"size":2768,"mtime":1607707601218,"results":"12","hashOfConfig":"8"},{"size":1246,"mtime":1607133044997,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"y00f6n",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/index.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/landing-page.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/workouts-list-sw.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/services/token-service-lf.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/services/auth-api-service-lf.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/header.js",[]]
+[{"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/App.js":"1","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/view-workout.js":"2","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Routes/login-route.js":"3","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/login-form.js":"4","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/header.js":"5","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/workout-list-item.js":"6","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Routes/sign-up-route.js":"7","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/sign-up-form.js":"8","/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/create-new-workout.js":"9"},{"size":3795,"mtime":1607797359215,"results":"10","hashOfConfig":"11"},{"size":2498,"mtime":1607797481834,"results":"12","hashOfConfig":"11"},{"size":748,"mtime":1607797325757,"results":"13","hashOfConfig":"11"},{"size":2031,"mtime":1607797447575,"results":"14","hashOfConfig":"11"},{"size":1247,"mtime":1607797465215,"results":"15","hashOfConfig":"11"},{"size":1327,"mtime":1607797497767,"results":"16","hashOfConfig":"11"},{"size":838,"mtime":1607797520626,"results":"17","hashOfConfig":"11"},{"size":3432,"mtime":1607797568018,"results":"18","hashOfConfig":"11"},{"size":4926,"mtime":1607797592490,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"y00f6n",{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/App.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/view-workout.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Routes/login-route.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/login-form.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/header.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/workout-list-item.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Routes/sign-up-route.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/sign-up-form.js",[],"/Users/sarawarnock/Projects/workout-creator-capstone-client/src/Components/create-new-workout.js",[]]
