+++ src/Components/navbar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import TokenService from './services/token-service';
+import TokenService from '../services/token-service';
 
 export default class Navbar extends React.Component {
   logOutClick = () => {
