+++ src/Components/view-past-workout.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import CircleButton from './CircleButton/circle-button'
 import { Link } from 'react-router-dom'
-import WorkOutContext from './context';
-import WorkoutApiService from './services/workout-api-service'
+import WorkOutContext from '../context';
+import WorkoutApiService from '../Services/workout-api-service'
 
 export default function ViewPastWorkout(props) {
   return (
