+++ src/Routes/public-route.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
-import TokenService from '../../Services/token-service';
-export default function PublicOnlyRoute({ component, ...props }) {
+import TokenService from '../Services/token-service-lf';
+
+
+export default function PublicRoute({ component, ...props }) {
   const Component = component;
   return (
     <Route
