--- src/log-in.js
@@ -1,180 +1,0 @@
-import React from 'react';
-import { Link  } from 'react-router-dom'
-import ValidationError from './validation-error'
-import AuthApiService from './services/auth-api-service'
-import TokenService from './services/token-service'
-
-export default class Login extends React.Component {
-  state = {
-    error: null,
-    loginEmail: {
-      value: '',
-      touched: false
-    },
-    loginPassword: {
-      value: '',
-      touched: false
-    },
-    errors: {
-      loginEmail: 'You must enter a valid username',
-      loginPassword: 'You must enter a valid password',
-    },
-    sessionUser: ''
-}
-
-  handleLoginSuccess = () => {
-    window.location = '/user/home'
-  }
-  
-  updateEmail(email) {
-    this.setState({ loginEmail: {value: email, touched: true } })
-  }
-
-  updatePassword(password) {
-    this.setState({ loginPassword: { value: password, touched: true } })
-  }
-
-  updateSessionUser(userId) {
-    this.setState({
-      sessionUser: userId
-    })
-  }
-
-  validateEmail(inputEmail) {
-    if (inputEmail == undefined) {
-       inputEmail = this.state.loginEmail.value.trim();
-    }
-    const mailFormat = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
-    if (!inputEmail.match(mailFormat)) {
-      return 'Valid username is required'
-    } return ' '
-  }
-
-  validatePassword(inputPassword) {
-    if (inputPassword == undefined) {
-      inputPassword = this.state.loginPassword.value.trim();
-   }
-    const passwordFormat = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])\w{8,}$/;
-    if (!inputPassword.match(passwordFormat)) {
-      return 'Valid password is required'
-    } return ' '
-  }
-
-  handleSubmit = (e) => {
-    e.preventDefault();
-    //create an object to store the search filters
-     const data = {}
-    // const { loginEmail, loginPassword } = e.target
-    
-
-    //get all the from data from the form component
-    const formData = new FormData(e.target)
-
-    //for each of the keys in form data populate it with form value
-    for (let value of formData) {
-        data[value[0]] = value[1]
-    }
-    console.log(data)
-    let {loginEmail, loginPassword} = data
-  
-    if (this.validateEmail(loginEmail) === '') {
-      this.setState({
-        error: 'You must enter a valid email'
-      })
-    }
-    if (this.validatePassword(loginPassword) === '') {
-      this.setState({
-        error: 'You must enter a valid password'
-      })
-    }
-    //assigning the object from the form data to params in the state
-    this.setState({
-        loginEmail: {
-          value: loginEmail
-        },
-        loginPassword: {
-          value: loginPassword
-        }
-    })
-
-    //check if the state is populated with the search params data
-    // console.log(this.state)
-
-    AuthApiService.postLogin({
-      email: loginEmail,
-      password: loginPassword
-    })
-    .then(response => {
-      if (response === undefined) {
-        console.log('empty response')
-        this.setState({
-          error: 'Email and/or password are not valid'
-        })
-        console.log(this.state)
-      }
-      else {
-        console.log('response ID', response)
-        TokenService.saveAuthToken(response.authToken) 
-        TokenService.saveUserId(response.userId)
-        TokenService.saveUserName(response.first_name)
-        window.location = `/home`
-        this.updateSessionUser(response.userId)
-        console.log(this.state.sessionUser)
-      }
-    })
-    .catch(err => {
-      console.log(err)
-      this.setState({
-        error: 'Email and/or password are not valid'
-      })
-    });
-  }
-
-  render() {
-    let validationError = ''
-    if (this.state.error != '') {
-      validationError = this.state.error
-    }
-    
-    const emailError = this.validateEmail();
-    const passwordError = this.validatePassword();
-    return (
-      <div className="App">
-        <main>
-            <h1>Log In</h1>
-                <form 
-                  className="login-form"
-                  onSubmit={this.handleSubmit}
-                >
-                  {validationError}
-                    <label className ="user-label" htmlFor="email">Username</label>
-                    <input 
-                      name="loginEmail"
-                      type="text" 
-                      id="email"
-                      onChange={e => this.updateEmail(e.target.value)}
-                      required
-                    />
-                    {this.state.loginEmail.touched && <ValidationError message={emailError} />}
-                    
-                    <label className ="user-label" htmlFor="password">Password</label>
-                    <input 
-                      name="loginPassword"
-                      type="password" 
-                      id="password"
-                      placeholder="Password"
-                      onChange={e => this.updatePassword(e.target.value)}
-                      required
-                    /> 
-                    {this.state.loginPassword.touched && <ValidationError message={passwordError} />}
-                    <button className="small-btn" type="submit"> Log In </button>
-                </form>
-                <div>
-                    <h3>Don't have an account yet?</h3>
-                    <Link to='/sign-up'><button className="small-btn">Sign Up</button></Link>
-                </div>
-        </main>
-      </div>
-    );
-  }
-}
